refactor: move mock data and use preview parameter
