+++ core/model/src/main/java/ksnd/hiraganaconverter/core/model/mock/MockConvertHistoryData.kt
@@ -1,9 +1,8 @@
-package ksnd.hiraganaconverter.feature.history
+package ksnd.hiraganaconverter.core.model.mock
 
 import ksnd.hiraganaconverter.core.model.ConvertHistoryData
 
-// TODO MockDataは:featureモジュールではなく別モジュールの方がよさそうだから移動する（:core:mockとか作ろうと思う）
-class MockConvertHistories {
+class MockConvertHistoryData {
     val data: List<ConvertHistoryData> = listOf(
         ConvertHistoryData(id = 1, before = "漢字", after = "かんじ", time = "2021/01/01 00:00:00"),
         ConvertHistoryData(id = 2, before = "漢字", after = "カンジ", time = "2021/01/01 00:00:00"),
